refactor: split server into multiple route files
