+++ src/server/index.ts
@@ -1,12 +1,7 @@
-import express, { Request, Response } from 'express';
+import express from 'express';
 import Rollbar from 'rollbar';
 import { config } from 'dotenv';
-import { home, math, notFound } from '../pages';
-import {
-  add, subtract, multiply, divide,
-} from '../js/utilities/arithmetic';
-import formatParams from './formatParams';
-import getErrorMessage from '../js/utilities/getErrorMessage';
+import { home, math, notFound } from './routes';
 
 config();
 
@@ -24,44 +19,11 @@ app.use(rollbar.errorHandler());
 
 app.use('/public', express.static('dist/public'));
 
-app.get('/', (request: Request, response: Response) => {
-  response.send(home());
-});
+app.get('/', home);
 
-app.get('/:operation/:a/:b', (request: Request, response: Response) => {
-  try {
-    let result = '';
-    const { operation, a, b } = formatParams(request.params);
+app.get('/:operation/:a/:b', math(rollbar));
 
-    switch (operation) {
-      case 'add':
-        result = `${a} + ${b} = ${add(a, b)}`;
-        break;
-      case 'subtract':
-        result = `${a} - ${b} = ${subtract(a, b)}`;
-        break;
-      case 'multiply':
-        result = `${a} &times; ${b} = ${multiply(a, b)}`;
-        break;
-      case 'divide':
-        result = `${a} &divide; ${b} = ${divide(a, b)}`;
-        break;
-      default:
-        break;
-    }
-    response.send(math(operation, result));
-  } catch (error) {
-    const errorMessage = getErrorMessage(error);
-    rollbar.error(errorMessage, request);
-
-    response.status(400).send(getErrorMessage(errorMessage));
-  }
-});
-
-app.get('*', (request: Request, response: Response) => {
-  rollbar.warn('Resource not found', request);
-  response.status(404).send(notFound());
-});
+app.get('*', notFound(rollbar));
 
 app.listen(PORT);
