+++ src/server/routes/index.ts
@@ -1,0 +1,3 @@
+export * from './home';
+export * from './math';
+export * from './404';
